Experimental turning code (description for deets)
Contains new code for turning the robot, all that's required to turn the robot is to press left or right on the joystick. If you press left the robot will move forward while turning left, the left wheels will spin at 1/4 speed of the wheels on the right. If you hold X while also holding left, the robot will move backward and turn. (I don't like how this works, I'd like to find an alternative)