Quality of life changes
New variables to change the division factor that's applied to motor speed when operating in the slow speed mode, and a variable to change the division factor that's applied to one motor when making a turn (so one side of the robot is able to move faster than the other). A new variable to change the time delayed when restarting from an emergency stop. Added comments to logic statements to make them easier to read. Cleaned up some redundant code.